+++ examples/resources/sonatyperepo_repository_swift_proxy/import.sh
@@ -1,0 +1,6 @@
+# Existing proxy Swift repository configuration can be imported as follows.
+#
+# NOTE: The Identifier REPOSITORY_NAME needs to match repository name in your Sonatype Nexus Repository instance.
+
+# Example
+terraform import sonatyperepo_repository_swift_proxy.example REPOSITORY_NAME